Initial Commit - Entities
In this INITIAL COMMIT, I am actually going by a design
document, and so have implemented everything in that
design document for the entity class and subclasses.
They are simple, as they do not need to be advanced for
this MVP, but they are clean. I feel like this is a
good first step for becoming an actual, genuine
programmer.

This has created a class for entity, player, and enemy,
and within entity are the general methods required by
all entities including getters/setters, basic level
and damage methods, and stat comparison methods.

The player and enemy classes for now just have basic
constructors with default parameters.