BATTLESCENE FIRST RUN
The Battlescene basics are now done. We have the most
simple case; two entities attack each other
automatically in the battlescene until one of them
dies, then the battlescene ends and the one who wins
is printed as winning and the one who loses is printed
as losing.

Much refactoring can still be done, and there still is
no proper takeTurn method. AttackEntity can also be
refactored and made simple with helpers in entity.

Checking when the battle is over can be refactored to
be cleaner and more complex.

However, from this simple case I will move to printing
things on the GUI.