GUI BEGINNINGS
This commit involves the very beginnings of using the
GUI. It is not clean code at all and doesn't actually
run the battle in the GUI, but it does append text and
do the stuff to create the GUI. Everything still runs
properly on the console. This has taken a while,
hence no commit the last two days. But this weekend
I will be committing a lot, hopefully.

It needs to be refactored, which will be later, but for
now I'm going to focus on actually bringing just the
outputs currently on the console into the GUI... w/o
mixing responsibilities too heinously between classes.

We'll see how that goes.