BATTLESCENE WORKING
In this commit, the BattleScene is working within the
GUI, with the addition of a controller. There isn't a
whole lot of functionality yet, but the gist of it is
there; the player can click attack and it properly
deals damage to the enemy.

The code needs to be refactored and full functionality
needs to be added, but this is definitely a start!