BATTLESCENE IN GUI WITH TURNS!
This commit involves the battle actually taking place
in the gui!!! Right now, the player can decide to
attack, use item, defend, or run, with only the first
actually doing something. It attacks the enemy until
the enemy is dead! But the enemy also attacks back
using the invokeLater situation. So, the battle works!

IT IS FUNCTIONAL! And it all prints in the GUI.

The only thing is adding the scrollbar always to be
at the bottom, and then of course refactoring, and
then I can work on adding features and such.