BUG FIXING: HEALING ATTACKS + TURN ORDER
Fixed (hopefully) two bugs this commit; first, that the
turn order wasn't working properly; the enemy would
attack over and over because there was no way to get
to player turn from enemy turn if determine first was
always non-player. Now it seems to work properly.

Secondly, fixed bug involving the attackEntity method,
now it takes the max of 0 or the difference to avoid
getting negatives which ADDED health to the enemy.