GAME PANEL + KINDA MESSED UP
This commit starts adding the infrastructure around and outside the battle scene by creating the game panel that will serve as the "shell" from which the
battle panel is created. Essentially, the game panel is an abstract class of all gui scenes that will take place (here, scenes will be referred to as
panels). The variables and fields as well as constructor are created, but the actual logic and situation needs to be added as well as how each subclass
will inherit from this abstract super.

Finally, I messed up a bit. I tried adding maven, then realized I really didn't want to have to deal with it, so now there are these .xml files in my
commit that I'm not sure if they're needed. I tried deleting them but everything broke, so I'm just going to commit them and hope things aren't broken...