TEXT LOG SEPARATED OUT
This commit creates a new class for the Text Log, so rather than only have it as a generic JTextArea in the GamePanel, it is an actual wrapper class of its own that can
contain all the necessary log and print methods. In here will be generic print methods that will be used in GamePanel and BattlePanel, hopefully with some better logic
to make them more readable, refactored, and generic. However, this commit just sets up the basic constructor and methods, and it is not yet implemented.