GAME PANEL FUNCTIONAL, BATTLE PANEL EXTENDS + TEXT LOG CLASS
This commit refactors a lot of GUI stuff, creating a separate class for the text log and moving many methods there. This also creates the generic game panel class, which will never be instantiated probably, but
instead serves as a super clas for the battle panel and then the story panel so that variables do not need to be repeated. Many methods were moved around, but the functionality appears to be the same throughout.
This will allow story panel and battle panel to share the same textLog, since they don't instantiate their own but rather use the one from GamePanel.

However, I have to check on that; if they extend GamePanel, and GamePanel creates a new textLog, wouldn't that still mean a new one is created every time the subclasses are created? I'll have to check on that.