BUTTON REORGANIZATION
This commit reorganizes how the buttons work. Rather than Action Listeners being created and set up in the BattleController, now there is an enum that takes in a label for the button and calls upon the
BattleController to do some action. When the battle panel (A new class) is created, it goes through the BattleButton enum and creates new buttons with those titles and actions, adding them to itself.

This does have one issue: the tighter coupling from the fact we need to pass the controller into the GUI. This might make things tricky, and maybe in future I will refactor this to make it less evil, but for now
it will work and saves me the hassle of having to define each button for each type of GUI scene we have. Now, I'm going to move on and do that for the other game gui things.