FOLDER REORG + NEW GUI
Moved some stuff again, and then created the story button panel and story gui for the actual out of battle gameplay. The only buttons will be yes or no (for now), and I want to have it where the story panel
shows "listening..." when those buttons are not in use. But that will all go in these new classes and panels! I also may rename the ButtonPanel class to ActionPanel, since it won't just have buttons but will
be the main way of interacting with the game (perhaps providing movement in a future re-run of this gui engine...)