BUG FIX + ADDED CLASS
Bug fix due to never actually adding a border layout to my gui panel, meaning the parts weren't being added to the layout correctly and the action panel would show up above the text log. That's now fixed.
The other thing I did was an add enum for GameChoice, which works for only yes or no questions. Something to be expanded in the future, but for now it's okay.

Another thing that this commit is demonstrating to me is that this has vastly escaped the original repo's original purpose of a battlescene MVP. This is turning much more into a Text-Based-Game Engine MVP. SO,
I'm going to try and move this branch over to a new repository just for that, and then use the battlescene mvp to actually flesh out the battle system more. What I'll probably do is de-bloat the version that
gets copied over.

I can either do this through intellij, or be really caveman about it and just copy and paste the files. Not sure which I'll do, but this will likely be the last commit on this branch.