VIEW MANAGER, SCENE CONTROLLER, GAMESTATE
This commit adds beginning infrastructure to the main crux of this project; being able to switch between scenes with a scene controller, being able to swap views with a view manager, and being
able to track the game with a game state. These are essentially the three main components of what this engine will do.

As a brief look into the design document in my head, the basic premise of this project is to:

Create a text-based java game engine utilizing the GUI that can swap between views, take player input to modify the game state and swap views, and go between battle scenes and story scenes.