ADDED SCENE CONTROLLER INHERITANCE
This commit adds the scene controller and its inheritance to the subclass controllers; MainMenuController and BattleController.

The super holds the single instance of the View Manager as well as the current GUI (and should likely hold the scene as well), and the others override that with their own data.

However, there is an issue with the fact that the GUI requires the controller to be called into it as a parameter. This is due to the way buttons are being done. So, I must move on to fixing the button
infrastructure and removing that tight coupling.