MAIN MENU REFACTORED
The main menu has been refactored to no longer need the controller and to allow a scene controller superclass between it and battle controller.

The main menu GUI no longer takes in a controller, and instead the addition of action listeners to the buttons are done through the controller itself, which I think makes sense; the view just
creates them and displays them, the controller gives them functionality. That might change one day but for now I think it works.