BATTLE GUI BEING REFACTORED
Battle GUI is being refactored in this commit, removing the references to the controller in the GUI and overall making the situation more readable and maintainable. The log has been moved around, as have the
references to the GUI and Button Panels. However, the printing methods need to be reworked, genericized, and made more readable rather than a long list of every method for every case.