PRINT GENERICIZED
The printing has roughly been genericized, or at least taken out of battle gui and made a bit more flexible. It's not great and involves a lot of switch cases, and can definitely be made better, but at least for
now it clears up that file. Plus, I wanted to commit when it was all done before any fixing had been done.

This does a lot, as a lot was involved in doing this sort of thing; adding some local variables, adding some abstract methods, and moving a lot of things into the GUI and GameplayGUI classes from
BattleGUI etc.

Speaking of fixing, there are some bugs here that need to be ironed out. One that I quickly fixed during this commit was an out-of-bounds error for attacking since I forgot the entity's name in the print health
method (even though I fixed it I think I accidentally put the wrong entity since I copy-pasted, so I'll need to fix that). There are other errors that arise including one when using an item with index 0 being
out-of-bounds for length 0, which is probably just some small oversight. Other than these, it seems to be working, and next commit will be fixing this bug.