STORY CONTROLLER CLASS AND MAIN MENU SWITCHING
This commit adds a story controller class to handle story scenes, or just out of battle gameplay scenes. AKA one of the three main scenes this MVP will handle.

It also adds the main menu controller with new game button functionality, which creates a new instance of the story GUI and immediately transitions the player there. There is a text log and will be a button panel
on this story gui, but that needs to be added in the next commit.