SCENE CHANGING!!!
Almost by accident, this prototype WORKS!!!!!!

The main menu starts, the new game button transitions to story mode which asks to start a battle, then the story mode transitions to the battle; all while keeping the same text log. Beyond that, the battle then
ends and goes back to the story mode, and if the player clicks yes, a new battle starts!

I want to add a system exit for clicking no, but after that... this is a working MVP ready for release!!!!

Well, sort of.

There is still the prickly issue that all this behavior is hard coded into the buttons, whereas I'd like to make those buttons generic and handle this separately. But that can be for a much later time. For now,
This is an actual environment in which a game can take place.... and that is awesome.