VIEW MANAGER, SCENE CONTROLLER, GAMESTATE (#1)
* VIEW MANAGER, SCENE CONTROLLER, GAMESTATE

This commit adds beginning infrastructure to the main crux of this project; being able to switch between scenes with a scene controller, being able to swap views with a view manager, and being
able to track the game with a game state. These are essentially the three main components of what this engine will do.

As a brief look into the design document in my head, the basic premise of this project is to:

Create a text-based java game engine utilizing the GUI that can swap between views, take player input to modify the game state and swap views, and go between battle scenes and story scenes.

* ADDED SCENE CONTROLLER INHERITANCE

This commit adds the scene controller and its inheritance to the subclass controllers; MainMenuController and BattleController.

The super holds the single instance of the View Manager as well as the current GUI (and should likely hold the scene as well), and the others override that with their own data.

However, there is an issue with the fact that the GUI requires the controller to be called into it as a parameter. This is due to the way buttons are being done. So, I must move on to fixing the button
infrastructure and removing that tight coupling.

* Refactoring buttons, again (#2)

* MAIN MENU REFACTORED

The main menu has been refactored to no longer need the controller and to allow a scene controller superclass between it and battle controller.

The main menu GUI no longer takes in a controller, and instead the addition of action listeners to the buttons are done through the controller itself, which I think makes sense; the view just
creates them and displays them, the controller gives them functionality. That might change one day but for now I think it works.

* BATTLE GUI BEING REFACTORED

Battle GUI is being refactored in this commit, removing the references to the controller in the GUI and overall making the situation more readable and maintainable. The log has been moved around, as have the
references to the GUI and Button Panels. However, the printing methods need to be reworked, genericized, and made more readable rather than a long list of every method for every case.

* PRINT GENERICIZED

The printing has roughly been genericized, or at least taken out of battle gui and made a bit more flexible. It's not great and involves a lot of switch cases, and can definitely be made better, but at least for
now it clears up that file. Plus, I wanted to commit when it was all done before any fixing had been done.

This does a lot, as a lot was involved in doing this sort of thing; adding some local variables, adding some abstract methods, and moving a lot of things into the GUI and GameplayGUI classes from
BattleGUI etc.

Speaking of fixing, there are some bugs here that need to be ironed out. One that I quickly fixed during this commit was an out-of-bounds error for attacking since I forgot the entity's name in the print health
method (even though I fixed it I think I accidentally put the wrong entity since I copy-pasted, so I'll need to fix that). There are other errors that arise including one when using an item with index 0 being
out-of-bounds for length 0, which is probably just some small oversight. Other than these, it seems to be working, and next commit will be fixing this bug.

* PRINT BUGFIXING

This commit fixes some bugs with printing, including an error when using an item and the wrong entity name printing when damage was taken and health printed. There still needs to be a lot of refactoring, but
printing is finally "genericized."

* STORY CONTROLLER CLASS AND MAIN MENU SWITCHING

This commit adds a story controller class to handle story scenes, or just out of battle gameplay scenes. AKA one of the three main scenes this MVP will handle.

It also adds the main menu controller with new game button functionality, which creates a new instance of the story GUI and immediately transitions the player there. There is a text log and will be a button panel
on this story gui, but that needs to be added in the next commit.

* SCENE CHANGING!!!

Almost by accident, this prototype WORKS!!!!!!

The main menu starts, the new game button transitions to story mode which asks to start a battle, then the story mode transitions to the battle; all while keeping the same text log. Beyond that, the battle then
ends and goes back to the story mode, and if the player clicks yes, a new battle starts!

I want to add a system exit for clicking no, but after that... this is a working MVP ready for release!!!!

Well, sort of.

There is still the prickly issue that all this behavior is hard coded into the buttons, whereas I'd like to make those buttons generic and handle this separately. But that can be for a much later time. For now,
This is an actual environment in which a game can take place.... and that is awesome.

* SYSTEM EXIT ON NO

Added a hardcoded value to the no button that exits the system with code 0. The only last touch would be adding a dialogue between battles asking if the player wants to continue, but whatever; that's just frills.

For now, this works, and the text log is preserved between all the screens. I can refactor this like hell and I really want to and I want to improve this, but I'm going to go back and improve some older projects
with all that I've learned from this one.

These last two mini projects have been really incredible. I've learned so much from just actually sitting down to think about what I want out of them beforehand rather than trying to basically make an entire game
out of them.

Now I can use these in the future, and combine them, as the battle scene will improve the battle code whereas this will focus on everything around that; GUI, buttons, etc.